pyproject.toml
[project]
name = "corridorkey"
version = "1.0.0"
description = "Neural network green screen keying for professional VFX pipelines"
readme = "README.md"
requires-python = ">=3.10, <3.14"
license = { text = "CC-BY-NC-SA-4.0" }
authors = [{ name = "Corridor Digital" }]

dependencies = [
    # Core inference
    "torch==2.8.0",
    "torchvision==0.23.0",
    "timm==1.0.24",
    "numpy",
    "opencv-python",
    "tqdm",
    "setuptools",
    # Triton fix for Windows
    # There might still be issues though https://github.com/triton-lang/triton-windows?tab=readme-ov-file#windows-file-path-length-limit-260-causes-compilation-failure
    "triton-windows==3.4.0.post21 ; sys_platform == 'win32'",
    # GVM alpha hint generator
    "diffusers",
    "transformers",
    "accelerate",
    "peft",
    "av",
    "Pillow",
    "PIMS",
    "easydict",
    "imageio",
    "matplotlib",
    # VideoMaMa alpha hint generator
    "einops",
    # BiRefNet alpha hint generator
    "kornia",
    # CLI tools (huggingface-hub is also a transitive dep, but must be direct
    # so that uv installs the "hf" console-script entry point)
    "huggingface-hub",
    # CLI framework + terminal output
    "typer>=0.12",
    "rich>=13",
]

[project.optional-dependencies]
cuda = [
    "torch==2.8.0",
    "torchvision==0.23.0",
]
mlx = [
    "corridorkey-mlx ; python_version >= '3.11'",
]
rocm = [
    "torch==2.8.0",
    "torchvision==0.23.0",
    "pytorch-triton-rocm==3.4.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
]
resolve = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
]

[dependency-groups]
dev = ["pytest", "pytest-cov", "ruff", "hypothesis"]
docs = ["zensical>=0.0.24"]

[project.scripts]
corridorkey = "corridorkey_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["CorridorKeyModule", "gvm_core", "VideoMaMaInferenceModule"]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = ["gvm_core/", "VideoMaMaInferenceModule/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]

[tool.ruff.lint.per-file-ignores]
# These files set environment variables (ROCm, OpenEXR) before importing
# torch/cv2 — the import order is intentional and must not be reordered.
"resolve_plugin/app.py" = ["E402"]
"resolve_plugin/engine_manager.py" = ["E402"]
"resolve_plugin/tests/conftest.py" = ["E402"]
"resolve_plugin/tests/test_*.py" = ["E402"]

[tool.ruff.format]
# Third-party code excluded via top-level exclude

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: requires CUDA GPU (skipped in CI)",
    "slow: long-running test",
    "mlx: requires Apple Silicon with MLX installed",
]
env = ["OPENCV_IO_ENABLE_OPENEXR=1", ]
addopts = "--tb=short"

[tool.coverage.run]
branch = true
# "." covers root-level modules (clip_manager.py, corridorkey_cli.py once PR #8 merges).
# Third-party and non-library code is excluded via omit.
source = ["CorridorKeyModule", "."]
omit = [
    "gvm_core/*",
    "VideoMaMaInferenceModule/*",
    "tests/*",
    "test_vram.py", # manual GPU smoke test, not part of the pytest suite
]

[tool.uv]
# Guard against transitive deps (diffusers, imageio, PIMS) silently pulling in
# opencv-python-headless, which conflicts with opencv-python at the file level
# (both install into the same cv2/ directory). If any future dep requests
# opencv-python-headless, uv resolution will fail explicitly rather than
# corrupting the environment.
constraint-dependencies = [
    "opencv-python-headless==99999",
]
conflicts = [
    [
        { extra = "cuda" },
        { extra = "mlx" },
        { extra = "rocm" },
    ],
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128" # CUDA 12.6 doesn't support RTX 5000 Series
explicit = true
extra = "cuda"

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.3"
explicit = true
extra = "rocm"

[tool.uv.sources]
# Use Hiera fix in order to utilize the FlashAttention Kernel
timm = { git = "https://github.com/Raiden129/pytorch-image-models-fix", branch = "fix/hiera-flash-attention-global-4d" }
torch = [
    { index = "pytorch", extra = "cuda" },
    { index = "pytorch-rocm", extra = "rocm" },
]
torchvision = [
    { index = "pytorch", extra = "cuda" },
    { index = "pytorch-rocm", extra = "rocm" },
]
pytorch-triton-rocm = { index = "pytorch-rocm", extra = "rocm" }
corridorkey-mlx = { git = "https://github.com/nikopueringer/corridorkey-mlx.git", extra = "mlx" }
